Changed how the graph data is loaded.
Graph data is now downloaded via CSV (same as before) but is now loaded
into an array and then feed into the graph. The next step would be to
update the graph with only new data without downloading the entire graph
data. I also changed some graph parameters that makes it work better
when dealing with a large amount of data.